Remove support for copy and move operations
Copy and move are very I/O heavy operations on an object storage system,
and should be avoided if not really required. These functions have been
removed for now.

The only operation that is still existing is to "rename" empty
containers and empty pseudo-folders. This is done by deleting the
container (which will fail if not empty) and recreating a new container
with the new name.